bg-remover-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
rembg==2.0.50
pillow==10.0.1
python-multipart==0.0.6
